Feat : 상품 상세 리포트 조회 API

+++ backend/gs-anomaly-detecion-backend/src/main/java/com/ssafy/s103/domain/anomalyproduct/application/repository/AnomalyProductRepository.java
@@ -4,6 +4,6 @@ import com.ssafy.s103.domain.anomalyproduct.entity.AnomalyProduct;
 import com.ssafy.s103.domain.anomalyproduct.entity.AnomalyStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnomalyProductRepository extends JpaRepository<AnomalyProduct, Integer>, AnomalyProductCustomRepository {
+public interface AnomalyProductRepository extends JpaRepository<AnomalyProduct, Long>, AnomalyProductCustomRepository {
     int countByStatus(AnomalyStatus status);
 }
